fix: restore pinned backend test contracts
