+++ packages/packaged-factories/factories/tts/factory.yaml
@@ -46,6 +46,9 @@ workers:
     modelProvider: CODEX
     modelLocality: LOCAL
     command: localai-vibevoice
+    args:
+      - --grpc-endpoint
+      - http://127.0.0.1:50051
     resources:
       - name: tts-cache
         capacity: 1
