+++ pkg/services/models/internal/backendconformance/published.go
@@ -1,3 +1,5 @@
+//go:build backendconformance || functionallong
+
 package backendconformance
 
 import (
