+++ pkg/services/models/internal/backendconformance/validation_test.go
@@ -1,3 +1,5 @@
+//go:build backendconformance
+
 package backendconformance
 
 import (
